OWGeneSets: new widget, schema only settings, minor improvements and code refactor
